--- Content.Shared/_RMC14/LinkAccount/SharedRMCGhostCosmetics.cs
@@ -1,10 +1,0 @@
-using Robust.Shared.Serialization;
-
-namespace Content.Shared._RMC14.LinkAccount;
-
-[Serializable, NetSerializable]
-public sealed record SharedRMCGhostCosmetics(
-    string? Particles,
-    string? Hat,
-    string? Mask
-);
