+++ Content.Shared/_RMC14/LinkAccount/SharedRMCPatronTier.cs
@@ -9,8 +9,6 @@ namespace Content.Shared._RMC14.LinkAccount;
 public sealed record SharedRMCPatronTier(
     bool ShowOnCredits,
     bool GhostColor,
-    bool GhostCosmetics, // Goob - ghost cosmetics
-    bool GhostParticles, // Goob - ghost cosmetics
     bool LobbyMessage,
     bool RoundEndShoutout,
     string Tier
