--- Resources/Locale/en-US/_Goobstation/ghost-cosmetics/ghost-cosmetics.ftl
@@ -1,54 +1,0 @@
-goob-ui-ghost-cosmetics = Ghost Cosmetics
-goob-ui-ghost-cosmetics-description = Dress up your ghost with a hat, a mask and a particle effect.
-goob-ui-ghost-cosmetics-particles = Particles
-goob-ui-ghost-cosmetics-hat = Hat
-goob-ui-ghost-cosmetics-mask = Mask
-goob-ui-ghost-cosmetics-none = None
-
-ghost-cosmetic-sparkles = Sparkles
-ghost-cosmetic-embers = Embers
-ghost-cosmetic-ectoplasm = Ectoplasm
-ghost-cosmetic-frost = Frost
-ghost-cosmetic-gloom = Gloom
-
-ghost-cosmetic-top-hat = Top Hat
-ghost-cosmetic-wizard-hat = Wizard Hat
-ghost-cosmetic-pirate-hat = Pirate Hat
-ghost-cosmetic-santa-hat = Santa Hat
-ghost-cosmetic-ushanka = Ushanka
-ghost-cosmetic-party-hat = Party Hat
-ghost-cosmetic-bowler-hat = Bowler Hat
-ghost-cosmetic-fedora = Fedora
-ghost-cosmetic-cat-ears = Cat Ears
-ghost-cosmetic-chef-hat = Chef's Hat
-ghost-cosmetic-cowboy-hat = Cowboy Hat
-ghost-cosmetic-dog-ears = Dog Ears
-ghost-cosmetic-fez = Fez
-ghost-cosmetic-jester-hat = Jester Hat
-ghost-cosmetic-paper-hat = Paper Hat
-ghost-cosmetic-plague-doctor-hat = Plague Doctor Hat
-ghost-cosmetic-sombrero = Sombrero
-ghost-cosmetic-straw-hat = Straw Hat
-ghost-cosmetic-witch-hat = Witch Hat
-ghost-cosmetic-christmas-crown = Christmas Crown
-
-ghost-cosmetic-clown-mask = Clown Mask
-ghost-cosmetic-mime-mask = Mime Mask
-ghost-cosmetic-fox-mask = Fox Mask
-ghost-cosmetic-bear-mask = Bear Mask
-ghost-cosmetic-golden-mask = Golden Mask
-ghost-cosmetic-gas-mask = Gas Mask
-ghost-cosmetic-bat-mask = Bat Mask
-ghost-cosmetic-bee-mask = Bee Mask
-ghost-cosmetic-breath-mask = Breath Mask
-ghost-cosmetic-cluwne-mask = Cluwne Mask
-ghost-cosmetic-moustache = Moustache
-ghost-cosmetic-jackal-mask = Jackal Mask
-ghost-cosmetic-joy-mask = Joy Mask
-ghost-cosmetic-muzzle = Muzzle
-ghost-cosmetic-neck-gaiter = Neck Gaiter
-ghost-cosmetic-ninja-mask = Ninja Mask
-ghost-cosmetic-plague-doctor-mask = Plague Doctor Mask
-ghost-cosmetic-rat-mask = Rat Mask
-ghost-cosmetic-raven-mask = Raven Mask
-ghost-cosmetic-sad-mime-mask = Sad Mime Mask
