+++ Resources/Prototypes/_Goobstation/Entities/Mobs/Customization/Markings/human_hair.yml
@@ -327,13 +327,13 @@
     - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
       state: nikooneshot
 
-- type: marking
-  id: HumanHairCotton
-  bodyPart: Hair
-  markingCategory: Hair
-  sprites:
-    - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
-      state: cotton
+#- type: marking CorvaxGoob-deleted
+#  id: HumanHairCotton
+#  bodyPart: Hair
+#  markingCategory: Hair
+#  sprites:
+#    - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
+#      state: cotton
 
 - type: marking
   id: HumanHairImpish
@@ -342,7 +342,7 @@
   sprites:
     - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
       state: impish
-      
+
 - type: marking
   id: HumanHairGloomyLonger
   bodyPart: Hair
@@ -350,7 +350,7 @@
   sprites:
     - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
       state: gloomyLonger
-      
+
 - type: marking
   id: HumanHairLila
   bodyPart: Hair
@@ -358,7 +358,7 @@
   sprites:
     - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
       state: lila
-      
+
 - type: marking
   id: HumanHairLongTails
   bodyPart: Hair
@@ -366,7 +366,7 @@
   sprites:
     - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
       state: longTails
-      
+
 - type: marking
   id: HumanHairStrictLong
   bodyPart: Hair
@@ -374,7 +374,7 @@
   sprites:
     - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
       state: strictLong
-      
+
 - type: marking
   id: HumanHairMermaid
   bodyPart: Hair
@@ -382,15 +382,15 @@
   sprites:
     - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
       state: mermaid
-      
-- type: marking
-  id: HumanHairShy
-  bodyPart: Hair
-  markingCategory: Hair
-  sprites:
-    - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
-      state: shy
-      
+
+#- type: marking CorvaxGoob-deleted
+#  id: HumanHairShy
+#  bodyPart: Hair
+#  markingCategory: Hair
+#  sprites:
+#    - sprite: _Goobstation/Mobs/Customization/human_hair.rsi
+#      state: shy
+
 - type: marking
   id: HumanHairTwinTail
   bodyPart: Hair
